+++ serial-version/file_reader.h
@@ -2,6 +2,6 @@
 #define EM_PROJECT_READER_SERIAL
 
 // reads from input file and populate the matrix of samples
-void fill_matrix(float* mat, int N, int D, char *FILE_PATH);
+void fill_matrix(double* mat, int N, int D, char *FILE_PATH);
 
 #endif
