+++ serial-version/utils.h
@@ -2,15 +2,15 @@
 #define EM_PROJECT_UTILS_SERIAL
 
 // calculate the inverse of matrix
-void inverse(float *A, float *inv, float *det, int n); // Function to calculate and store inverse
+void inverse(double *A, double *inv, double *det, int n); // Function to calculate and store inverse
 
 // calculate matrix-vector multiplication
-void matmul(float *mat, float *vec, float *res, int D);
+void matmul(double *mat, double *vec, double *res, int D);
 
 // calculate the dot product
-float dotProduct(float *a, float *b, int D);
+double dotProduct(double *a, double *b, int D);
 
 // standardize the training data
-void standardize(float* data, int N, int D);
+void standardize(double* data, int N, int D);
 
 #endif
